Suggest adding a reference link to the document
We've had a few customers ask how to deal with the issue in the NOTE section below. I suggest we link to the doc that explains how to move the ASA and SPN when in coexistence. This section talks about how you cannot share the ASA, but it does not link to the document that tells you how to correct it. This causes customers to do this incorrectly and therefore causing critical connectivity issues. 

Suggest change this section 
Important

Exchange 2010 and Exchange 2016 can't share the same ASA credential. If your ASA credential was created for Exchange 2010, you have to create a new one for Exchange 2016.


To this: 
 Important

Exchange 2010 and Exchange 2016 can't share the same ASA credential. If your ASA credential was created for Exchange 2010, you have to create a new one for Exchange 2016. For more information about creating a new ASA credential and removing HTTP SPN from Exchange 2010 ASA, please see https://techcommunity.microsoft.com/t5/exchange-team-blog/exchange-2013-and-exchange-2010-coexistence-with-kerberos/ba-p/611196

Or link however you see fit.